'semi-finish'

+++ controllers/videoController.js
@@ -1,5 +1,6 @@
 import video from '../models/video';
 import routes from '../routes';
+import Comment from '../models/comment';
 
 export const home = async (req, res) => {
   //async and await
@@ -42,8 +43,11 @@ export const postUpload = async (req, res) => {
   const newVideo = await video.create({
     fileUrl: path,
     title,
-    description
+    description,
+    creator: req.user.id
   });
+  req.user.videos.push(newVideo.id);
+  req.user.save();
   res.redirect(routes.videoDetail(newVideo.id));
 };
 
@@ -53,7 +57,16 @@ export const videoDetail = async (req, res) => {
   } = req;
 
   try {
-    const tempVideo = await video.findById(id);
+    const tempVideo = await video
+      .findById(id)
+      .populate('creator')
+      .populate({
+        path: 'comments',
+        populate: {
+          path: 'creator'
+        }
+      });
+    console.log(tempVideo);
     res.render('videoDetail', { pageTitle: tempVideo.title, video: tempVideo });
   } catch (error) {
     console.log(error);
@@ -67,6 +80,9 @@ export const editVideo = async (req, res) => {
   } = req;
   try {
     const tempVideo = await video.findById(id);
+    if (tempVideo.creator !== req.user.id) {
+      throw Error();
+    }
     res.render('editVideo', {
       pageTitle: `Edit ${tempVideo.title}`,
       video: tempVideo
@@ -101,9 +117,54 @@ export const deleteVideo = async (req, res) => {
   } = req;
 
   try {
+    const tempVideo = await video.findById(id);
+    if (tempVideo.creator !== req.user.id) {
+      throw Error();
+    }
     await video.findOneAndRemove({ _id: id });
   } catch (error) {
     console.log(error);
   }
   res.redirect(routes.home);
+};
+
+//api
+
+export const registerView = async (req, res) => {
+  const {
+    params: { id }
+  } = req;
+  try {
+    const tempVideo = await video.findById(id);
+    tempVideo.views += 1;
+    tempVideo.save();
+    res.status(200);
+  } catch (error) {
+    res.status(400);
+  } finally {
+    res.end();
+  }
+};
+
+export const addComment = async (req, res) => {
+  const {
+    params: { id },
+    body: { comment },
+    user
+  } = req;
+
+  try {
+    const tempVideo = await video.findById(id);
+    const newComment = await Comment.create({
+      text: comment,
+      creator: user.id
+    });
+
+    tempVideo.comments.push(newComment.id);
+    tempVideo.save();
+  } catch (error) {
+    res.status(400);
+  } finally {
+    res.end();
+  }
 };
